docs: improve public documentation (#1083)
* feat: update googletest module

* docs: improve docs

* docs: improve docs v2

* docs: add missing docs

* fix: fix compilation warnings

+++ tests/modules/commerce_test.cpp
@@ -54,6 +54,7 @@ TEST_F(CommerceTest, shouldGenerateSkuWithSpecifiedLength)
                                                                    { return skuCharacter == numericCharacter; });
                                     }));
 }
+
 TEST_F(CommerceTest, shouldGenerateProductFullName)
 {
     const auto generatedProductFullName = productFullName();
@@ -65,18 +66,23 @@ TEST_F(CommerceTest, shouldGenerateProductFullName)
     {
         for (size_t j = i + 1; j < words.size(); ++j)
         {
-            const std::string adjective = common::join(std::vector<std::string_view>(words.begin(), words.begin() + i), " ");
-            const std::string material = common::join(std::vector<std::string_view>(words.begin() + i, words.begin() + j), " ");
-            const std::string name     = common::join(std::vector<std::string_view>(words.begin() + j, words.end()), " ");
+            const std::string adjective = common::join(
+                std::vector<std::string_view>(words.begin(), words.begin() + static_cast<std::ptrdiff_t>(i)), " ");
+            const std::string material =
+                common::join(std::vector<std::string_view>(words.begin() + static_cast<std::ptrdiff_t>(i),
+                                                           words.begin() + static_cast<std::ptrdiff_t>(j)),
+                             " ");
+            const std::string name = common::join(
+                std::vector<std::string_view>(words.begin() + static_cast<std::ptrdiff_t>(j), words.end()), " ");
 
             const bool adjectiveMatch = std::ranges::any_of(productAdjectives, [&adjective](const std::string_view& a)
                                                             { return a == adjective; });
 
-            const bool materialMatch = std::ranges::any_of(productMaterials, [&material](const std::string_view& m)
-                                                           { return m == material; });
+            const bool materialMatch =
+                std::ranges::any_of(productMaterials, [&material](const std::string_view& m) { return m == material; });
 
-            const bool nameMatch = std::ranges::any_of(productNames, [&name](const std::string_view& n)
-                                                       { return n == name; });
+            const bool nameMatch =
+                std::ranges::any_of(productNames, [&name](const std::string_view& n) { return n == name; });
 
             if (adjectiveMatch && materialMatch && nameMatch)
             {
@@ -88,9 +94,9 @@ TEST_F(CommerceTest, shouldGenerateProductFullName)
             break;
     }
 
-    ASSERT_TRUE(isValid) << "Generated product full name \"" << generatedProductFullName << "\" is not a valid combination.";
+    ASSERT_TRUE(isValid) << "Generated product full name \"" << generatedProductFullName
+                         << "\" is not a valid combination.";
 }
-
 
 TEST_F(CommerceTest, shouldGenerateProductAdjective)
 {
